requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
apsw==3.51.0.0
apswutils==0.1.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
distro==1.9.0
fastcore==1.8.16
fastlite==0.2.1
frozenlist==1.8.0
groq==0.36.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
itsdangerous==2.2.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-groq==1.1.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.48
marshmallow==3.26.1
multidict==6.7.0
mypy-extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
propcache==0.4.1
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-fasthtml==0.12.35
python-multipart==0.0.20
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
starlette==0.50.0
tavily==1.1.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
